refactor: move responsibilities of BlockGetter and State supervision
