+++ apps/omg_watcher/lib/application.ex
@@ -17,147 +17,102 @@ defmodule OMG.Watcher.Application do
   use Application
   use OMG.API.LoggerExt
 
-  defmodule WatcherSupervisor do
-    @moduledoc """
-    This supervisor takes care of starting (and keeping running) all modules of which Watcher consists of
-    """
+  def start(_type, _args) do
+    import Supervisor.Spec
 
-    use Supervisor
+    # Define workers and child supervisors to be supervised
+    block_finality_margin = Application.get_env(:omg_api, :ethereum_event_block_finality_margin)
+    slow_exit_validator_block_margin = Application.get_env(:omg_watcher, :slow_exit_validator_block_margin)
 
-    defmodule ChainStateSupervisor do
-      @moduledoc """
-      This supervisor takes care of BlockGetter and State processes.
-      In case one process fails, this supervisor's role is to restore consistent state
-      """
-      use Supervisor
-
-      def start_link do
-        Supervisor.start_link(__MODULE__, :ok)
-      end
-
-      def init(:ok) do
-        # State and Block Getter are linked, because they must restore their state to the last stored state
-        # If Block Getter fails, it starts from the last checkpoint while State might have had executed
-        # some transactions. Such a situation will cause error when trying to execute already executed transaction.
-        children = [
-          {OMG.API.State, []},
+    children = [
+      # Start the Ecto repository
+      supervisor(OMG.Watcher.DB.Repo, []),
+      # Start workers
+      {OMG.Watcher.Eventer, []},
+      {OMG.API.RootChainCoordinator, MapSet.new([:depositer, :fast_validator, :slow_validator, :block_getter])},
+      worker(
+        OMG.API.EthereumEventListener,
+        [
           %{
-            id: :block_getter,
-            start: {OMG.Watcher.BlockGetter, :start_link, [[]]},
-            restart: :transient
+            synced_height_update_key: :last_depositor_eth_height,
+            service_name: :depositer,
+            block_finality_margin: block_finality_margin,
+            get_events_callback: &OMG.Eth.RootChain.get_deposits/2,
+            process_events_callback: &deposit_events_callback/1,
+            get_last_synced_height_callback: &OMG.DB.last_depositor_eth_height/0
           }
-        ]
+        ],
+        id: :depositer
+      ),
+      worker(
+        OMG.API.EthereumEventListener,
+        [
+          %{
+            block_finality_margin: 0,
+            synced_height_update_key: :last_fast_exit_eth_height,
+            service_name: :fast_validator,
+            get_events_callback: &OMG.Eth.RootChain.get_exits/2,
+            process_events_callback: &exit_events_callback/1,
+            get_last_synced_height_callback: &OMG.DB.last_fast_exit_eth_height/0
+          }
+        ],
+        id: :fast_validator
+      ),
+      worker(
+        OMG.API.EthereumEventListener,
+        [
+          %{
+            block_finality_margin: slow_exit_validator_block_margin,
+            synced_height_update_key: :last_slow_exit_eth_height,
+            service_name: :slow_validator,
+            get_events_callback: &OMG.Eth.RootChain.get_exits/2,
+            process_events_callback:
+              OMG.Watcher.ExitValidator.Validator.challenge_invalid_exits(&slow_validator_utxo_exists_callback/1),
+            get_last_synced_height_callback: &OMG.DB.last_slow_exit_eth_height/0
+          }
+        ],
+        id: :slow_validator
+      ),
+      # Start the endpoint when the application starts
+      supervisor(OMG.Watcher.Web.Endpoint, []),
+      # Start BlockGetter and State
+      supervisor(OMG.Watcher.BlockGetter.Supervisor, [])
+    ]
 
-        opts = [strategy: :one_for_all, name: __MODULE__]
-        Supervisor.init(children, opts)
-      end
-    end
+    _ = Logger.info(fn -> "Started application OMG.Watcher.Application" end)
 
-    def start_link do
-      Supervisor.start_link(__MODULE__, :ok)
-    end
+    # See https://hexdocs.pm/elixir/Supervisor.html
+    # for other strategies and supported options
+    opts = [strategy: :one_for_one, name: OMG.Watcher.Supervisor]
+    Supervisor.start_link(children, opts)
+  end
 
-    def init(:ok) do
-      import Supervisor.Spec
+  # Tell Phoenix to update the endpoint configuration
+  # whenever the application is updated.
+  def config_change(changed, _new, removed) do
+    OMG.Watcher.Web.Endpoint.config_change(changed, removed)
+    :ok
+  end
 
-      # Define workers and child supervisors to be supervised
-      block_finality_margin = Application.get_env(:omg_api, :ethereum_event_block_finality_margin)
-      slow_exit_validator_block_margin = Application.get_env(:omg_watcher, :slow_exit_validator_block_margin)
-
-      children = [
-        # Start the Ecto repository
-        supervisor(OMG.Watcher.DB.Repo, []),
-        # Start workers
-        {OMG.Watcher.Eventer, []},
-        {OMG.API.RootChainCoordinator, MapSet.new([:depositer, :fast_validator, :slow_validator, :block_getter])},
-        worker(
-          OMG.API.EthereumEventListener,
-          [
-            %{
-              synced_height_update_key: :last_depositor_eth_height,
-              service_name: :depositer,
-              block_finality_margin: block_finality_margin,
-              get_events_callback: &OMG.Eth.RootChain.get_deposits/2,
-              process_events_callback: &deposit_events_callback/1,
-              get_last_synced_height_callback: &OMG.DB.last_depositor_eth_height/0
-            }
-          ],
-          id: :depositer
-        ),
-        worker(
-          OMG.API.EthereumEventListener,
-          [
-            %{
-              block_finality_margin: 0,
-              synced_height_update_key: :last_fast_exit_eth_height,
-              service_name: :fast_validator,
-              get_events_callback: &OMG.Eth.RootChain.get_exits/2,
-              process_events_callback: &exit_events_callback/1,
-              get_last_synced_height_callback: &OMG.DB.last_fast_exit_eth_height/0
-            }
-          ],
-          id: :fast_validator
-        ),
-        worker(
-          OMG.API.EthereumEventListener,
-          [
-            %{
-              block_finality_margin: slow_exit_validator_block_margin,
-              synced_height_update_key: :last_slow_exit_eth_height,
-              service_name: :slow_validator,
-              get_events_callback: &OMG.Eth.RootChain.get_exits/2,
-              process_events_callback:
-                OMG.Watcher.ExitValidator.Validator.challenge_invalid_exits(&slow_validator_utxo_exists_callback/1),
-              get_last_synced_height_callback: &OMG.DB.last_slow_exit_eth_height/0
-            }
-          ],
-          id: :slow_validator
-        ),
-        # Start the endpoint when the application starts
-        supervisor(OMG.Watcher.Web.Endpoint, []),
-        # Start Block Getter and State
-        supervisor(__MODULE__.ChainStateSupervisor, [])
-      ]
-
-      _ = Logger.info(fn -> "Started application OMG.Watcher.Application" end)
-
-      # See https://hexdocs.pm/elixir/Supervisor.html
-      # for other strategies and supported options
-      opts = [strategy: :one_for_one, name: OMG.Watcher.Supervisor.Unlinked]
-      Supervisor.init(children, opts)
-    end
-
-    # Tell Phoenix to update the endpoint configuration
-    # whenever the application is updated.
-    def config_change(changed, _new, removed) do
-      OMG.Watcher.Web.Endpoint.config_change(changed, removed)
+  defp slow_validator_utxo_exists_callback(utxo_exit) do
+    with :ok <- OMG.API.State.exit_if_not_spent(utxo_exit) do
       :ok
-    end
-
-    defp slow_validator_utxo_exists_callback(utxo_exit) do
-      with :ok <- OMG.API.State.exit_if_not_spent(utxo_exit) do
-        :ok
-      else
-        :utxo_does_not_exist ->
-          :ok = OMG.Watcher.ChainExiter.exit()
-          :child_chain_exit
-      end
-    end
-
-    defp deposit_events_callback(deposits) do
-      :ok = OMG.API.State.deposit(deposits)
-      _ = OMG.Watcher.DB.EthEvent.insert_deposits(deposits)
-      :ok
-    end
-
-    defp exit_events_callback(exits) do
-      :ok = OMG.Watcher.ExitValidator.Validator.challenge_invalid_exits(fn _ -> :ok end).(exits)
-      _ = OMG.Watcher.DB.EthEvent.insert_exits(exits)
-      :ok
+    else
+      :utxo_does_not_exist ->
+        :ok = OMG.Watcher.ChainExiter.exit()
+        :child_chain_exit
     end
   end
 
-  def start(_type, _args) do
-    WatcherSupervisor.start_link()
+  defp deposit_events_callback(deposits) do
+    :ok = OMG.API.State.deposit(deposits)
+    _ = OMG.Watcher.DB.EthEvent.insert_deposits(deposits)
+    :ok
+  end
+
+  defp exit_events_callback(exits) do
+    :ok = OMG.Watcher.ExitValidator.Validator.challenge_invalid_exits(fn _ -> :ok end).(exits)
+    _ = OMG.Watcher.DB.EthEvent.insert_exits(exits)
+    :ok
   end
 end
